+++ gears/carto_gears_api/lib/carto_gears_api/gear_controller_helper.rb
@@ -7,5 +7,9 @@ module CartoGearsApi
     include TrackjsHelper
     include GoogleAnalyticsHelper
     include HubspotHelper
+    include FrontendConfigHelper
+    include AppAssetsHelper
+    include MapsApiHelper
+    include SqlApiHelper
   end
 end
